implement tabs for challenges, teams and users
